Funciona todo!

+++ tpoDatos/src/main/java/com/example/uade/tpoDatos/service/CarritoService.java
@@ -4,11 +4,12 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import com.example.uade.tpoDatos.controllers.UserController;
 import com.example.uade.tpoDatos.entity.Carrito;
+import com.example.uade.tpoDatos.entity.CarritoUsuario;
 import com.example.uade.tpoDatos.repository.CarritoRepository;
 import com.example.uade.tpoDatos.repository.ProductoRepository;
 import com.example.uade.tpoDatos.service.interfaces.CarritoServiceImpl;
@@ -23,28 +24,36 @@ public class CarritoService implements CarritoServiceImpl{
     private ProductoRepository productoRepository;
 
     @Override
-    public ResponseEntity<String> agregarACarrito(Carrito carrito) {
-        Carrito carritoExsistente = carritoRepository.findByNombreProducto(carrito.getProducto());
-
-        if(carritoExsistente == null){
+    public ResponseEntity<String> agregarACarrito(CarritoUsuario carritoUsuario) {
+        
+        Carrito carritoExistente = carritoRepository.findByNombreProducto(carritoUsuario.getProducto());
+        
+        if(carritoExistente == null){
             try {
-                carritoRepository.save(carrito);
+                Carrito carritoNuevo = new Carrito();
+                carritoNuevo.setId(carritoUsuario.getId());
+                carritoNuevo.setCantidad(carritoUsuario.getCantidad());
+                carritoNuevo.setProducto(carritoUsuario.getProducto());
+                carritoNuevo.setNombreUsuario(UserController.getUsuarioActivo().getNombre());
+                carritoRepository.save(carritoNuevo);
                 return new ResponseEntity<String>("Producto agregado al carrito!", HttpStatus.CREATED);
             } catch (Exception e) {
                 return new ResponseEntity<>("Error al guardar el producto en el carrito", HttpStatus.INTERNAL_SERVER_ERROR);
             }
         }
         else{
-            if(productoRepository.findByNombre(carrito.getProducto()).equals(carrito)){
+            if(productoRepository.findByNombre(carritoExistente.getProducto()) != null){
                 try {
-                    carritoExsistente.setCantidad(carritoExsistente.getCantidad() + carrito.getCantidad());
-                    carritoRepository.save(carritoExsistente);
+                    carritoExistente.setCantidad(carritoExistente.getCantidad() + carritoUsuario.getCantidad());
+                    carritoRepository.save(carritoExistente);
                     return new ResponseEntity<String>("Cantidad de producto actualizada",HttpStatus.ACCEPTED);
                 } catch (Exception e) {
                     return new ResponseEntity<String>("Error al actualizar la cantidad", HttpStatus.INTERNAL_SERVER_ERROR);
                 }
             }
-            return new ResponseEntity<String>("El producto no existe", HttpStatus.BAD_REQUEST);
+            else{
+                return new ResponseEntity<String>("El producto no existe", HttpStatus.BAD_REQUEST);
+            }
         }
     }
 
@@ -52,7 +61,7 @@ public class CarritoService implements CarritoServiceImpl{
     public ResponseEntity<String> eliminarDeCarrito(String nombreProducto) {
         List<Carrito> elementosCarrito = carritoRepository.findAll();
         for(Carrito c: elementosCarrito){
-            if(c.getProducto().equals(nombreProducto)){
+            if(c.getNombreUsuario().equals(UserController.getUsuarioActivo().getNombre()) && c.getProducto().equals(nombreProducto)){
                 try {
                     carritoRepository.delete(c);
                     return new ResponseEntity<String>("Producto eliminado del carrito con exito", HttpStatus.ACCEPTED);
